Replace old affinity check
