+++ cscs-checks/prgenv/affinity_check.py
@@ -3,15 +3,27 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
+
 import os
+
 import reframe as rfm
 import reframe.utility.sanity as sn
+import reframe.utility.osext as osext
+from reframe.core.exceptions import SanityError
 
 
 class AffinityTestBase(rfm.RegressionTest):
-    def __init__(self, variant):
+    '''Base class for the affinity checks.
+
+    It reads a reference file for each valid system, which allows this base
+    class to figure out the processor's topology. The content of this reference
+    file is simply the output of the `lscpu -e` command. For more info on this,
+    see the `read_proc_topo` hook.
+    '''
+
+    def __init__(self):
         self.valid_systems = ['daint:gpu', 'daint:mc',
-                              'dom:gpu', 'dom:mc']
+                              'dom:gpu', 'dom:mc', 'eiger:mc']
         self.valid_prog_environs = ['PrgEnv-gnu']
         self.build_system = 'Make'
         self.build_system.options = ['-C affinity', 'MPI=1']
@@ -21,166 +33,521 @@ class AffinityTestBase(rfm.RegressionTest):
         self.sourcesdir = os.path.join('src/affinity_ref')
         self.prebuild_cmds = ['git clone https://github.com/vkarak/affinity']
         self.executable = './affinity/affinity'
-        self.variant = variant
+        self.sanity_patterns = self.assert_consumed_cpu_set()
         self.maintainers = ['RS', 'SK']
         self.tags = {'production', 'scs', 'maintenance', 'craype'}
 
-    @rfm.run_before('sanity')
-    def set_sanity(self):
+        # Dict with the partition's topology - output of "lscpu -e"
+        self.topology = {
+            'dom:gpu':   'topo_dom_gpu.txt',
+            'dom:mc':    'topo_dom_mc.txt',
+            'daint:gpu': 'topo_dom_gpu.txt',
+            'daint:mc':  'topo_dom_mc.txt',
+            'eiger:mc':  'topo_eiger_mc.txt',
+        }
+
+    @rfm.run_before('compile')
+    def set_topo_file(self):
+        cp = self.current_partition.fullname
+        self.topo_file = self.topology[cp]
+
+    @rfm.run_after('compile')
+    def read_proc_topo(self):
+        '''Import the processor's topology from the reference file.
+
+        This hook inserts the following attributes based on the processor's
+        topology:
+            - cpu_set: set containing all the cpu IDs
+            - num_cpus
+            - num_cpus_per_core
+            - num_numa_nodes
+            - num_sockets
+            - numa_nodes: dictionary containing the cpu sets for each numa
+                node. The keys of the dictionary are simply the ID of the
+                numa node.
+            - sockets: dictionary containing the cpu sets for each socket.
+                The keys of the dictionary are simply the socket IDs.
+
+        This hook requires the reference file, so the earliest it can run is
+        after the compilation stage, once the required files have been copied
+        over to the stage directory.
+        '''
+        cp = self.current_partition.fullname
+        with osext.change_dir(self.stagedir):
+            lscpu = sn.extractall(
+                r'^\s*(\d+)\s*(\d+)\s*(\d+)\s*(\d+)',
+                self.topology[cp], 0,
+                lambda x: [int(xi) for xi in x.split()]
+            ).evaluate()
+
+        # Build the cpu set
+        self.cpu_set = set(map(lambda x: x[0], lscpu))
+        self.num_cpus = len(self.cpu_set)
+
+        # Build the numa sets
+        self.num_numa_nodes = len(set(map(lambda x: x[1], lscpu)))
+        self.num_cpus_per_core = int(
+            self.num_cpus/len(set(map(lambda x: x[3], lscpu)))
+        )
+        self.numa_nodes = {}
+        for i in range(self.num_numa_nodes):
+            self.numa_nodes[i] = set(
+                map(lambda y: y[0], filter(lambda x: x[1] == i, lscpu))
+            )
+
+        # Build the socket sets
+        self.num_sockets = len(set(map(lambda x: x[2], lscpu)))
+        self.sockets = {}
+        for i in range(self.num_sockets):
+            self.sockets[i] = set(
+                map(lambda y: y[0], filter(lambda x: x[2] == i, lscpu))
+            )
+
+        # Store the lscpu output
+        self._lscpu = lscpu
+
+    def get_sibiling_cpus(self, cpuid, by=None):
+        '''Return a cpu set where cpuid belongs to.
+
+        The cpu set can be extracted by matching core, numa domain or socket.
+        This is controlled by the `by` argument.
+        '''
+        _map = {
+            'core': 3,
+            'socket': 2,
+            'numa': 1,
+        }
+
+        if (cpuid < 0) or (cpuid >= self.num_cpus):
+            raise TaskExit(f'a cpuid with value {cpuid} is invalid')
+
+        if by is None:
+            raise TaskExit('must specify the sibiling level')
+        else:
+            try:
+                sibiling_level = _map[by]
+            except KeyError:
+                raise TaskExit('invalid sibiling level')
+
+        sibiling_id = list(
+            filter(lambda x: x[0] == cpuid, self._lscpu)
+        )[0][sibiling_level]
+
+        return set(
+            map(
+                lambda y: y[0],
+                filter(
+                    lambda x: x[sibiling_level] == sibiling_id,
+                    self._lscpu
+                )
+            )
+        )
+
+    @sn.sanity_function
+    def assert_consumed_cpu_set(self):
+        '''Check that all the resources have been consumed.
+
+        Tests derived from this class must implement a hook that consumes
+        the cpu set as the results from the affinity tool are processed.
+        '''
+        return sn.assert_eq(self.cpu_set, set())
+
+    @rfm.run_after('run')
+    def parse_output(self):
+        '''Extract the data from the affinity tool.'''
+        re_aff_cpus = r'CPU affinity: \[\s+(?P<cpus>[\d+\s+]+)\]'
+        re_aff_thrds = r'^Tag:[^\n\r]*Thread:\s+(?P<thread>\d+)'
+        re_aff_ranks = r'^Tag:[^\n\r]*Rank:\s+(?P<rank>\d+)[\s+\S+]'
 
         def parse_cpus(x):
             return sorted([int(xi) for xi in x.split()])
 
-        re_aff_cores = r'CPU affinity: \[\s+(?P<cpus>[\d+\s+]+)\]'
-        self.aff_cores = sn.extractall(
-            re_aff_cores, self.stdout, 'cpus', parse_cpus)
-        ref_key = 'ref_' + self.current_partition.fullname
-        self.ref_cores = sn.extractall(
-            re_aff_cores, self.cases[self.variant][ref_key],
-            'cpus', parse_cpus)
-        re_aff_thrds = r'^Tag:[^\n\r]*Thread:\s+(?P<thread>\d+)'
-        self.aff_thrds = sn.extractall(re_aff_thrds, self.stdout, 'thread',
-                                       int)
-        self.ref_thrds = sn.extractall(
-            re_aff_thrds, self.cases[self.variant][ref_key],
-            'thread', int)
-        re_aff_ranks = r'^Tag:[^\n\r]*Rank:\s+(?P<rank>\d+)[\s+\S+]'
-        self.aff_ranks = sn.extractall(re_aff_ranks, self.stdout, 'rank', int)
-        self.ref_ranks = sn.extractall(
-            re_aff_ranks, self.cases[self.variant][ref_key],
-            'rank', int)
+        with osext.change_dir(self.stagedir):
+            self.aff_cpus = sn.extractall(
+                re_aff_cpus, self.stdout, 'cpus', parse_cpus
+            ).evaluate()
 
-        # Ranks and threads can be extracted into lists in order to compare
-        # them since the affinity programm prints them in ascending order.
-        self.sanity_patterns = sn.all([
-            sn.assert_eq(self.aff_thrds, self.ref_thrds),
-            sn.assert_eq(self.aff_ranks, self.ref_ranks),
-            sn.assert_eq(sn.sorted(self.aff_cores), sn.sorted(self.ref_cores))
-        ])
+            self.aff_thrds = sn.extractall(
+                re_aff_thrds, self.stdout, 'thread', int
+            )
+
+            self.aff_ranks = sn.extractall(
+                re_aff_ranks, self.stdout, 'rank', int
+            )
 
     @rfm.run_before('run')
     def set_multithreading(self):
-        self.use_multithreading = self.cases[self.variant]['multithreading']
-
-
-@rfm.parameterized_test(['omp_bind_threads'],
-                        ['omp_bind_threads_nomultithread'],
-                        ['omp_bind_cores'])
-class AffinityOpenMPTest(AffinityTestBase):
-    def __init__(self, variant):
-        super().__init__(variant)
-        self.descr = 'Checking the cpu affinity for OMP threads'
-        self.cases = {
-            'omp_bind_threads': {
-                'ref_daint:gpu': 'gpu_omp_bind_threads.txt',
-                'ref_dom:gpu': 'gpu_omp_bind_threads.txt',
-                'ref_daint:mc': 'mc_omp_bind_threads.txt',
-                'ref_dom:mc': 'mc_omp_bind_threads.txt',
-                'num_cpus_per_task:gpu': 24,
-                'num_cpus_per_task:mc': 72,
-                'ntasks_per_core': 2,
-                'multithreading': None,
-                'OMP_PLACES': 'threads',
-            },
-            'omp_bind_threads_nomultithread': {
-                'ref_daint:gpu': 'gpu_omp_bind_threads_nomultithread.txt',
-                'ref_dom:gpu': 'gpu_omp_bind_threads_nomultithread.txt',
-                'ref_daint:mc': 'mc_omp_bind_threads_nomultithread.txt',
-                'ref_dom:mc': 'mc_omp_bind_threads_nomultithread.txt',
-                'num_cpus_per_task:gpu': 12,
-                'num_cpus_per_task:mc': 36,
-                'ntasks_per_core': None,
-                # When `--hint=nomultithread` is not explicitly expecified only
-                # half of the physical cores are used.
-                'multithreading': False,
-                'OMP_PLACES': 'threads',
-            },
-            'omp_bind_cores': {
-                'ref_daint:gpu': 'gpu_omp_bind_cores.txt',
-                'ref_dom:gpu': 'gpu_omp_bind_cores.txt',
-                'ref_daint:mc': 'mc_omp_bind_cores.txt',
-                'ref_dom:mc': 'mc_omp_bind_cores.txt',
-                'num_cpus_per_task:gpu': 12,
-                'num_cpus_per_task:mc': 36,
-                'ntasks_per_core': 1,
-                'multithreading': None,
-                'OMP_PLACES': 'cores',
-            },
-        }
-        self.variant = variant
-
-    @rfm.run_before('run')
-    def set_tasks_per_core(self):
-        partname = self.current_partition.name
-        self.num_cpus_per_task = (
-            self.cases[self.variant]['num_cpus_per_task:%s' % partname])
-        if self.cases[self.variant]['ntasks_per_core']:
-            self.num_tasks_per_core = (
-                self.cases[self.variant]['ntasks_per_core'])
-
-        self.num_tasks = 1
-        self.variables = {
-            'OMP_NUM_THREADS': str(self.num_cpus_per_task),
-            'OMP_PLACES': self.cases[self.variant]['OMP_PLACES']
-            # OMP_PROC_BIND is set to TRUE if OMP_PLACES is defined.
-            # Both OMP_PROC_BIND values CLOSE and SPREAD give the same
-            # result as OMP_PROC_BIND=TRUE when all cores are requested.
-        }
-
-
-@rfm.parameterized_test(['alternate_socket_filling'],
-                        ['consecutive_socket_filling'],
-                        ['single_task_per_socket_omp'],
-                        ['single_task_per_socket_omp_multithread'])
-class SocketDistributionTest(AffinityTestBase):
-    def __init__(self, variant):
-        super().__init__(variant)
-        self.descr = 'Checking distribution of ranks and threads over sockets.'
-        self.valid_systems = ['daint:mc', 'dom:mc']
-        self.cases = {
-            'alternate_socket_filling': {
-                'ref_daint:mc': 'alternate_socket_filling.txt',
-                'ref_dom:mc': 'alternate_socket_filling.txt',
-                'num_tasks': 36,
-                'num_cpus_per_task': 1,
-                'num_tasks_per_socket': 18,
-                'multithreading': False,
-                'cpu-bind': None,
-            },
-            'consecutive_socket_filling': {
-                'ref_daint:mc': 'consecutive_socket_filling.txt',
-                'ref_dom:mc': 'consecutive_socket_filling.txt',
-                'num_tasks': 36,
-                'num_cpus_per_task': 1,
-                'num_tasks_per_socket': None,
-                'multithreading': False,
-                'cpu-bind': 'rank',
-            },
-            'single_task_per_socket_omp': {
-                'ref_daint:mc': 'single_task_per_socket_omp.txt',
-                'ref_dom:mc': 'single_task_per_socket_omp.txt',
-                'num_tasks': 2,
-                'num_cpus_per_task': 18,
-                'num_tasks_per_socket': 1,
-                'multithreading': False,
-                'cpu-bind': None,
-            },
-            'single_task_per_socket_omp_multithread': {
-                'ref_daint:mc': 'single_task_per_socket_omp_multithread.txt',
-                'ref_dom:mc': 'single_task_per_socket_omp_multithread.txt',
-                'num_tasks': 2,
-                'num_cpus_per_task': 36,
-                'num_tasks_per_socket': 1,
-                'multithreading': True,
-                'cpu-bind': None,
-            },
-        }
-        self.num_tasks = self.cases[variant]['num_tasks']
-        self.num_cpus_per_task = self.cases[variant]['num_cpus_per_task']
-        self.num_tasks_per_socket = self.cases[variant]['num_tasks_per_socket']
+        '''Hook to control multithreading settings for each system.'''
+        cp = self.current_partition.fullname
+        mthread = self.system.get(cp, {}).get('multithreading', None)
+        if mthread:
+            self.use_multithreading = mthread
 
     @rfm.run_before('run')
     def set_launcher(self):
-        if self.cases[self.variant]['cpu-bind']:
-            self.job.launcher.options = ['--cpu-bind=%s' %
-                                         self.cases[self.variant]['cpu-bind']]
+        '''Hook to control hints and cpu-bind for each system.'''
+        cp = self.current_partition.fullname
+        cpu_bind = self.system.get(cp, {}).get('cpu-bind', None)
+        if cpu_bind:
+            self.job.launcher.options += [f'--cpu-bind={cpu_bind}']
+
+        hint = self.system.get(cp, {}).get('hint', None)
+        if hint:
+            self.job.launcher.options += [f'--hint={hint}']
+
+
+class AffinityOpenMPBase(AffinityTestBase):
+    '''Extend affinity base with OMP hooks.
+
+    The tests derived from this class book the full node and place the
+    threads acordingly based exclusively on the OMP_BIND env var. The
+    number of total OMP_THREADS will vary depending on what are we
+    binding the OMP threads to (e.g. if we bind to sockets, we'll have as
+    many threads as sockets).
+    '''
+
+    # FIXME: PR #1699 should make these vars instead of parameters.
+    parameter('omp_bind')
+    parameter('omp_proc_bind', ['spread'])
+
+    def __init__(self):
+        super().__init__()
+        self.num_tasks = 1
+
+    @property
+    def ncpus_per_task(self):
+        '''We use this property to set the hook below and keep exec order.'''
+        return self.num_cpus
+
+    @rfm.run_before('run')
+    def set_num_cpus_per_task(self):
+        self.num_cpus_per_task = self.ncpus_per_task
+
+    @rfm.run_before('run')
+    def set_omp_vars(self):
+        self.variables = {
+            'OMP_NUM_THREADS': str(self.num_omp_threads),
+            'OMP_PLACES': self.omp_bind,
+            'OMP_PROC_BIND': self.omp_proc_bind,
+        }
+
+    @rfm.run_before('sanity')
+    def consume_cpu_set(self):
+        raise ValueError('this function must be overridden')
+
+
+@rfm.simple_test
+class OneTheadPerCPUOpenMP(AffinityOpenMPBase):
+    '''Pin each OMP thread to a different CPU.'''
+    parameter('omp_bind', ['threads'])
+
+    def __init__(self):
+        super().__init__()
+        self.descr = 'Pin one OMP thread per CPU.'
+
+        # System-dependent settings here
+        self.system = {}
+
+    @property
+    def num_omp_threads(self):
+        # One OMP thread per cpu
+        return self.num_cpus
+
+    @rfm.run_before('sanity')
+    def consume_cpu_set(self):
+        '''Threads are bound to cpus.'''
+        for affinity_set in self.aff_cpus:
+            # Affinity set must be of length 1 and CPU ID cannot be repeated
+            if ((len(affinity_set) > 1) or not all(x in self.cpu_set
+                                                   for x in affinity_set)):
+                # This will force the sanity function to fail.
+                raise SanityError('incorrect affinity set')
+
+            # Decrement the affinity set from the cpu set
+            self.cpu_set -= set(affinity_set)
+
+
+@rfm.simple_test
+class OneThreadPerCoreOpenMP(AffinityOpenMPBase):
+    '''Pin each OMP thread to a different core.'''
+    parameter('omp_bind', ['cores'])
+
+    def __init__(self):
+        super().__init__()
+        self.descr = 'Pin one OMP thread per core.'
+
+        # System-dependent settings here
+        self.system = {}
+
+    @property
+    def num_omp_threads(self):
+        # One OMP thread per core
+        return int(self.num_cpus/self.num_cpus_per_core)
+
+    @rfm.run_before('sanity')
+    def consume_cpu_set(self):
+        '''Threads are bound to cores.'''
+        for affinity_set in self.aff_cpus:
+
+            # Get CPU sibilings by core
+            cpu_sibilings = self.get_sibiling_cpus(affinity_set[0], by='core')
+
+            # All CPUs in the set must belong to the same core
+            if (not all(x in self.cpu_set for x in affinity_set) or
+                not all(x in cpu_sibilings for x in affinity_set)):
+                raise SanityError('incorrect affinity set')
+
+            # Decrement the cpu set with all the CPUs that belong to this core
+            self.cpu_set -= cpu_sibilings
+
+
+@rfm.simple_test
+class OneThreadPerCoreOpenMPnomultithread(OneThreadPerCoreOpenMP):
+    '''Only one cpu per core booked.'''
+
+    def __init__(self):
+        super().__init__()
+        self.descr = 'Pin one OMP thread per core.'
+        self.use_multithreading = False
+
+        # System-dependent settings here
+        self.system = {}
+
+    @property
+    def ncpus_per_task(self):
+        return self.num_omp_threads
+
+    @rfm.run_before('sanity')
+    def assert_aff_set_length(self):
+        '''Only 1 CPU pinned per thread.'''
+        if not all(len(aff_set) == 1 for aff_set in self.aff_cpus):
+            raise SanityError('incorrect affinity set')
+
+
+@rfm.simple_test
+class OneThreadPerSocketOpenMP(AffinityOpenMPBase):
+    '''Pin each OMP thread to a different socket.'''
+    parameter('omp_bind', ['socket'])
+
+    def __init__(self):
+        super().__init__()
+        self.descr = 'Pin one OMP thread per socket.'
+
+        # System-dependent settings here
+        self.system = {}
+
+    @property
+    def num_omp_threads(self):
+        # One OMP thread per core
+        return self.num_sockets
+
+    @rfm.run_before('sanity')
+    def consume_cpu_set(self):
+        '''Threads are bound to sockets.'''
+        for affinity_set in self.aff_cpus:
+
+            # Get CPU sibilings by socket
+            cpu_sibilings = self.get_sibiling_cpus(
+                affinity_set[0], by='socket')
+
+            # Alll CPUs in the affinity set must belong to the same socket
+            if (not all(x in self.cpu_set for x in affinity_set) or
+                not all(x in cpu_sibilings for x in affinity_set)):
+                raise SanityError('incorrect affinity set')
+
+            # Decrement all the CPUs in this socket from the cpu set.
+            self.cpu_set -= cpu_sibilings
+
+
+@rfm.simple_test
+class OneTaskPerSocketOpenMPnomultithread(AffinityOpenMPBase):
+    '''One task per socket, and as many OMP threads as cores per socket.'''
+    parameter('omp_bind', ['sockets'])
+    parameter('omp_proc_bind', ['close'])
+
+    def __init__(self):
+        super().__init__()
+        self.descr = 'One task per socket - wo. multithreading.'
+        self.use_multithreading = False
+
+        # System-dependent settings here
+        self.system = {}
+
+    @property
+    def num_omp_threads(self):
+        return int(self.num_cpus/self.num_cpus_per_core/self.num_sockets)
+
+    @property
+    def ncpus_per_task(self):
+        return self.num_omp_threads
+
+    @rfm.run_before('run')
+    def set_num_tasks(self):
+        self.num_tasks = self.num_sockets
+
+    @rfm.run_before('sanity')
+    def consume_cpu_set(self):
+
+        threads_in_socket = [0]*self.num_sockets
+
+        def get_socket_id(cpuid):
+            for i in range(self.num_sockets):
+                if cpuid in self.sockets[i]:
+                    return i
+
+        for affinity_set in self.aff_cpus:
+            # Count the number of OMP threads that live on each socket
+            threads_in_socket[get_socket_id(affinity_set[0])] += 1
+
+            # Get CPU sibilings by socket
+            cpu_sibilings = self.get_sibiling_cpus(
+                affinity_set[0], by='socket'
+            )
+
+            # The size of the affinity set matches the number of OMP threads
+            # and all CPUs from the set belong to the same socket.
+            if ((self.num_omp_threads != len(affinity_set)) or
+                not all(x in cpu_sibilings for x in affinity_set)):
+                raise SanityError('incorrect affinity set')
+
+        # Remove the sockets the cpu set.
+        for i, socket in enumerate(self.sockets.values()):
+            if threads_in_socket[i] == self.num_omp_threads:
+                self.cpu_set -= socket
+
+
+@rfm.simple_test
+class OneTaskPerSocketOpenMP(OneTaskPerSocketOpenMPnomultithread):
+    '''One task per socket, and as many OMP threads as CPUs per socket.
+
+    We can reuse the test above. Just need to change the multithreading flag
+    and the number of OMP threads.
+    '''
+
+    def __init__(self):
+        super().__init__()
+        self.descr = 'One task per socket - w. multithreading.'
+        self.use_multithreading = True
+
+        # System-dependent settings here
+        self.system = {}
+
+    @property
+    def num_omp_threads(self):
+        return int(self.num_cpus/self.num_sockets)
+
+
+@rfm.simple_test
+class ConsecutiveSocketFilling(AffinityTestBase):
+    '''Fill the sockets with the tasks in consecutive order.
+
+    This test uses as many tasks as physical cores available in a node.
+    Multithreading is disabled.
+    '''
+
+    def __init__(self):
+        super().__init__()
+        self.use_multithreading = False
+
+        # System-dependent settings
+        self.system = {}
+
+    @rfm.run_before('run')
+    def set_tasks(self):
+        self.num_tasks = int(self.num_cpus/self.num_cpus_per_core)
+        self.num_cpus_per_task = 1
+        self.job.launcher.options += [f'--cpu-bind=rank']
+
+    @rfm.run_before('sanity')
+    def consume_cpu_set(self):
+        '''Check that all physical cores have been used in the right order.'''
+        task_count = 0
+        for socket_number in range(self.num_sockets):
+            # Keep track of the CPUs present in this socket
+            cpus_present = set()
+            for task_number in range(int(self.num_tasks/self.num_sockets)):
+                # Get the list of CPUs with affinity
+                affinity_set = self.aff_cpus[task_count]
+
+                # Only 1 CPU per affinity set allowed.
+                if (len(affinity_set) > 1) or (any(cpu in cpus_present
+                                                   for cpu in affinity_set)):
+                    raise SanityError(
+                        f'incorrect affinity set for task {task_count}'
+                    )
+
+                else:
+                    cpus_present.update(
+                        self.get_sibiling_cpus(affinity_set[0], by='core')
+                    )
+
+                task_count += 1
+
+            # Ensure all CPUs belong to the same socket
+            cpuset_by_socket = self.get_sibiling_cpus(
+                next(iter(cpus_present)), by='socket'
+            )
+            if not all(cpu in cpuset_by_socket for cpu in cpus_present):
+                raise SanityError(
+                    f'socket {socket_number} not filled in order'
+                )
+
+            else:
+                # Decrement the current socket from the available CPU set
+                self.cpu_set -= cpus_present
+
+
+@rfm.simple_test
+class AlternateSocketFilling(AffinityTestBase):
+    '''Sockets are filled in a round-robin fashion.
+
+    This test uses as many tasks as physical cores available in a node.
+    Multithreading is disabled.
+    '''
+
+    def __init__(self):
+        super().__init__()
+        self.use_multithreading = False
+        self.system = {
+            # System-dependent settings here
+        }
+
+    @rfm.run_before('run')
+    def set_tasks(self):
+        self.num_tasks = int(self.num_cpus/self.num_cpus_per_core)
+        self.num_cpus_per_task = 1
+        self.num_tasks_per_socket = int(self.num_tasks/self.num_sockets)
+
+    @rfm.run_before('sanity')
+    def consume_cpu_set(self):
+        '''Check that consecutive tasks are round-robin pinned to sockets.'''
+
+        # Get a set per socket to keep track of the CPUs
+        sockets = [set() for s in range(self.num_sockets)]
+        task_count = 0
+        for task in range(self.num_tasks_per_socket):
+            for s in range(self.num_sockets):
+                # Get the list of CPUs with affinity
+                affinity_set = self.aff_cpus[task_count]
+
+                # Only 1 CPU per affinity set is allowed
+                if (len(affinity_set) > 1) or (any(cpu in sockets[s]
+                                                   for cpu in affinity_set)):
+                    raise SanityError(
+                        f'incorrect affinity set for task {task_count}'
+                    )
+
+                else:
+                    sockets[s].update(
+                        self.get_sibiling_cpus(affinity_set[0], by='core')
+                    )
+
+                task_count += 1
+
+            # Check that all sockets have the same CPU count
+            if not all(len(s) == (task+1)*2 for s in sockets):
+                self.cpu_set.add(-1)
+
+        # Decrement the sockets from the CPU set
+        for s in sockets:
+            self.cpu_set -= s
