--- cscs-checks/prgenv/src/affinity_ref/alternate_socket_filling.txt
@@ -1,72 +1,0 @@
-Tag: affinity/affinity, Hostname: nid00404, Rank: 0, Thread: 0
-    CPU affinity: [   0]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 1, Thread: 0
-    CPU affinity: [  18]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 2, Thread: 0
-    CPU affinity: [   1]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 3, Thread: 0
-    CPU affinity: [  19]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 4, Thread: 0
-    CPU affinity: [   2]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 5, Thread: 0
-    CPU affinity: [  20]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 6, Thread: 0
-    CPU affinity: [   3]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 7, Thread: 0
-    CPU affinity: [  21]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 8, Thread: 0
-    CPU affinity: [   4]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 9, Thread: 0
-    CPU affinity: [  22]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 10, Thread: 0
-    CPU affinity: [   5]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 11, Thread: 0
-    CPU affinity: [  23]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 12, Thread: 0
-    CPU affinity: [   6]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 13, Thread: 0
-    CPU affinity: [  24]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 14, Thread: 0
-    CPU affinity: [   7]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 15, Thread: 0
-    CPU affinity: [  25]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 16, Thread: 0
-    CPU affinity: [   8]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 17, Thread: 0
-    CPU affinity: [  26]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 18, Thread: 0
-    CPU affinity: [   9]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 19, Thread: 0
-    CPU affinity: [  27]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 20, Thread: 0
-    CPU affinity: [  10]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 21, Thread: 0
-    CPU affinity: [  28]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 22, Thread: 0
-    CPU affinity: [  11]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 23, Thread: 0
-    CPU affinity: [  29]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 24, Thread: 0
-    CPU affinity: [  12]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 25, Thread: 0
-    CPU affinity: [  30]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 26, Thread: 0
-    CPU affinity: [  13]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 27, Thread: 0
-    CPU affinity: [  31]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 28, Thread: 0
-    CPU affinity: [  14]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 29, Thread: 0
-    CPU affinity: [  32]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 30, Thread: 0
-    CPU affinity: [  15]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 31, Thread: 0
-    CPU affinity: [  33]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 32, Thread: 0
-    CPU affinity: [  16]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 33, Thread: 0
-    CPU affinity: [  34]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 34, Thread: 0
-    CPU affinity: [  17]
-Tag: affinity/affinity, Hostname: nid00404, Rank: 35, Thread: 0
-    CPU affinity: [  35]
