--- cscs-checks/prgenv/src/affinity_ref/gpu_omp_bind_threads.txt
@@ -1,48 +1,0 @@
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 0
-    CPU affinity: [   0]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 1
-    CPU affinity: [   1]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 2
-    CPU affinity: [   2]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 3
-    CPU affinity: [   3]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 4
-    CPU affinity: [   4]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 5
-    CPU affinity: [   5]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 6
-    CPU affinity: [   6]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 7
-    CPU affinity: [   7]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 8
-    CPU affinity: [   8]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 9
-    CPU affinity: [   9]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 10
-    CPU affinity: [  10]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 11
-    CPU affinity: [  11]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 12
-    CPU affinity: [  12]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 13
-    CPU affinity: [  13]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 14
-    CPU affinity: [  14]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 15
-    CPU affinity: [  15]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 16
-    CPU affinity: [  16]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 17
-    CPU affinity: [  17]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 18
-    CPU affinity: [  18]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 19
-    CPU affinity: [  19]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 20
-    CPU affinity: [  20]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 21
-    CPU affinity: [  21]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 22
-    CPU affinity: [  22]
-Tag: affinity/affinity, Hostname: nid00001, Rank: 0, Thread: 23
-    CPU affinity: [  23]
